% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up_data.R
\name{clean_up_data}
\alias{clean_up_data}
\alias{clean_up_data.data.frame}
\alias{clean_up_data.list}
\title{Clean up experiment data}
\usage{
clean_up_data(.data, ...)

\method{clean_up_data}{data.frame}(.data, subjprefix = "Exp-")

\method{clean_up_data}{list}(l, ..., remove_gk = TRUE, .w = TRUE)
}
\arguments{
\item{subjprefix}{the prefix you want to use for shortening `WorkerId`s}

\item{remove_gk}{boolean, to remove the gatekeeper from Exp4}

\item{.w}{boolean, set to `FALSE` if you don't want a warning}

\item{df}{a data frame}
}
\description{
Right now, applies only to Experiment 1, 2, 3, and 4.
}
