% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up_data.R
\name{collapse_exp4}
\alias{collapse_exp4}
\title{Turn the Exp4 list results into a data frame}
\usage{
collapse_exp4(l, ..., remove_gk = TRUE)
}
\arguments{
\item{l}{the Exp4 list}

\item{\dots}{unused, but important to make `remove_gk` obligatorily named}

\item{remove_gk}{a boolean, to remove the gatekeeper blocks in the result}
}
\description{
The results for Exp4 in the package come as a list. This turns the results into a data frame. Be wary that it joins by 'Condition' by default.
}
