% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprime.R
\name{dprime}
\alias{dprime}
\title{Calculate d'}
\usage{
dprime(n_hit, n_miss, n_fa, n_cr)
}
\arguments{
\item{n_hit}{hits = true positives}

\item{n_miss}{misses = false negatives}

\item{n_fa}{false alarms = false positives}

\item{n_cr}{correctly rejected = true negatives}
}
\description{
Calculates d'
}
