% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{exclusion_step}
\alias{exclusion_step}
\title{Get exclusion steps for each experiment}
\usage{
exclusion_step(exp = c("Exp1", "Exp2", "Exp3"), exc_step = 1)
}
\arguments{
\item{exp}{A string of the experiment name}

\item{exc_step}{An integer of the exclusion step}
}
\value{
A function that takes in a dataframe and outputs a list of `WorkerId`s
}
\description{
Given an experiment and a numeric step value, returns a function that takes a data frame and outputs a list of `WorkerId`s (not guarunteed to be unique) for input into the function producted by `start_exclusions()`.
}
\examples{
exclude_exp1 <- start_exclusions(Condition)
d.ex1_goodsubj <- d.ex1_raw \%>\%
  {exclude_exp1(., "Technical problem", exclusion_step("Exp1", 1)(.))} \%>\%
  {exclude_exp1(., "> 4 visual catch trial false negatives", exclusion_step("Exp1", 2)(.))} \%>\%
  {exclude_exp1(., "> 15 visual catch trial false positives", exclusion_step("Exp1", 2)(.))} \%>\%
  {exclude_exp1(., "Took a break > 10 min", exclusion_step("Exp1", 4)(.))} \%>\%
  {exclude_exp1(., "Multilingual", exclusion_step("Exp1", 5)(.))} \%>\%
  {exclude_exp1(., "Did not wear headphones", exclusion_step("Exp1", 6)(.))} \%>\%
  {exclude_exp1(., "Did not take experiment in quiet location", exclusion_step("Exp1", 7)(.))}

}
