% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{exp4_gk_exclusion_crutch}
\alias{exp4_gk_exclusion_crutch}
\title{A hack for making the Exp4 GK exclusions easier}
\usage{
exp4_gk_exclusion_crutch(l, excl_fn, ...)
}
\description{
A hack for making the Exp4 GK exclusions easier
}
