% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{mark_mistakes}
\alias{mark_mistakes}
\title{Mark catch trial mistakes}
\usage{
mark_mistakes(df, slow_presses = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{slow_presses}{a boolean}
}
\description{
This takes in a data frame and marks false negative and false positive catch trials, by default accepting 'slow' button presses
}
\section{Slow button presses}{


When `slow_presses` is `TRUE`, we do the following: for catch trials, we simply ask whether space was pressed on the current or next trial-=-if yes, it's correct; for non-catch trials, we ask whether space was not pressed or if space was pressed and the previous trial was a catch trial---if yes, it's correct; everything else is a mistake.
}

