% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check_norm.R
\name{sanity_check_exp4}
\alias{sanity_check_exp4}
\title{Does a few sanity checks on the Exp 4 data}
\usage{
sanity_check_exp4(l)
}
\arguments{
\item{l}{The list of data frames}
}
\value{
Nothing
}
\description{
Just for 'Exp 4'
}
