% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{start_exclusions}
\alias{start_exclusions}
\title{Exclude and keep track of participants}
\usage{
start_exclusions(..., is_exp4 = FALSE)
}
\arguments{
\item{\dots}{The names of any additional columns to keep in the excluded subjects dataframe. Should have constant values within subjects.}
}
\value{
A function (described below)
}
\description{
A tidy way of excluding and keeping track of exlcuded participants.
}
\section{The return function}{


The return value of `start_exclusions` is a function that has the following arguments:

- `df`: The dataframe you want to exclude participants from
- `reason`: The stated reason for these participants' exclusion

- \dots: Either the input to a `filter()` function that selects the participants to be excluded, or a list of WorkerIds to be excluded

- `just_ids=TRUE`: If you're supplying a list of filtering criteria, this should be `FALSE`

- `give_df=FALSE`: If true, returns the excluded participants dataframe (and doesn't do any excluding)
}

