% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unipluck}
\alias{unipluck}
\title{`pluck` uniquely, more easily}
\usage{
unipluck(df, col)
}
\arguments{
\item{df}{a data frame}

\item{col}{a column name}
}
\description{
Basically just `pluck()` for data frames, but it turns bare columns into strings and then gets the unique values.
}
